# openapi.json
{"openapi":"3.1.0","info":{"title":"Pi-PM","description":"Personal Intelligence Portfolio Manager","version":"0.4.1"},"paths":{"/api/v1/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Api V1 Health Get"}}}}}}},"/api/v1/stocks":{"get":{"tags":["stocks"],"summary":"List Stocks","operationId":"list_stocks_api_v1_stocks_get","parameters":[{"name":"data_status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockRead"},"title":"Response List Stocks Api V1 Stocks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stocks/{symbol}":{"get":{"tags":["stocks"],"summary":"Get Stock","operationId":"get_stock_api_v1_stocks__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stocks/{symbol}/market-data":{"get":{"tags":["stocks"],"summary":"Get Stock Market Data","operationId":"get_stock_market_data_api_v1_stocks__symbol__market_data_get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"}},{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":5000,"minimum":1},{"type":"null"}],"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketDataRead"},"title":"Response Get Stock Market Data Api V1 Stocks  Symbol  Market Data Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market-data/ingest":{"post":{"tags":["market-data"],"summary":"Ingest Market Data","operationId":"ingest_market_data_api_v1_market_data_ingest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketDataIngestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketDataIngestResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rankings/run":{"post":{"tags":["rankings"],"summary":"Run Ranking","operationId":"run_ranking_api_v1_rankings_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RankingRunRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RankingRunRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rankings/latest":{"get":{"tags":["rankings"],"summary":"Get Latest Ranking","operationId":"get_latest_ranking_api_v1_rankings_latest_get","parameters":[{"name":"universe_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"}},{"name":"strategy_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"}},{"name":"strategy_version","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RankingRunRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rankings/{run_id}":{"get":{"tags":["rankings"],"summary":"Get Ranking Run","operationId":"get_ranking_run_api_v1_rankings__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RankingRunRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rankings/{run_id}/top":{"get":{"tags":["rankings"],"summary":"Get Ranking Top","operationId":"get_ranking_top_api_v1_rankings__run_id__top_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Run Id"}},{"name":"n","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"N"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RankingTopRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtest/generate-rankings":{"post":{"tags":["backtest"],"summary":"Generate Rankings","operationId":"generate_rankings_api_v1_backtest_generate_rankings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRankingsRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateRankingsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/backtest/summary":{"get":{"tags":["backtest"],"summary":"Get Backtest Summary","operationId":"get_backtest_summary_api_v1_backtest_summary_get","parameters":[{"name":"start_date","in":"query","required":true,"schema":{"type":"string","format":"date","title":"Start Date"}},{"name":"end_date","in":"query","required":true,"schema":{"type":"string","format":"date","title":"End Date"}},{"name":"universe_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"}},{"name":"strategy_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"}},{"name":"strategy_version","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BacktestSummaryRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/validation/backfill":{"post":{"tags":["validation"],"summary":"Backfill Validation","operationId":"backfill_validation_api_v1_validation_backfill_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationBackfillRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationBackfillResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/validation/runs/{run_id}/compute":{"post":{"tags":["validation"],"summary":"Compute Validation","operationId":"compute_validation_api_v1_validation_runs__run_id__compute_post","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Run Id"}},{"name":"force_recompute","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Force Recompute"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationReportRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/validation/runs/{run_id}":{"get":{"tags":["validation"],"summary":"Get Validation Report","operationId":"get_validation_report_api_v1_validation_runs__run_id__get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationReportRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/validation/runs/{run_id}/snapshots":{"get":{"tags":["validation"],"summary":"Get Validation Snapshots","operationId":"get_validation_snapshots_api_v1_validation_runs__run_id__snapshots_get","parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ValidationSnapshotRead"},"title":"Response Get Validation Snapshots Api V1 Validation Runs  Run Id  Snapshots Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/validation/summary":{"get":{"tags":["validation"],"summary":"Get Validation Summary","operationId":"get_validation_summary_api_v1_validation_summary_get","parameters":[{"name":"universe_code","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"}},{"name":"strategy_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"}},{"name":"strategy_version","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"}},{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date"}},{"name":"horizon","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Horizon"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationSummaryRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BacktestSummaryRead":{"properties":{"universe_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"},"strategy_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"},"strategy_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"},"start_date":{"type":"string","title":"Start Date"},"end_date":{"type":"string","title":"End Date"},"ranking_runs_total":{"type":"integer","title":"Ranking Runs Total"},"validated_runs_total":{"type":"integer","title":"Validated Runs Total"},"pending_validation_runs":{"type":"integer","title":"Pending Validation Runs"}},"type":"object","required":["start_date","end_date","ranking_runs_total","validated_runs_total","pending_validation_runs"],"title":"BacktestSummaryRead"},"GenerateRankingsRequest":{"properties":{"universe_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"strategy_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"},"strategy_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"},"benchmark_symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Benchmark Symbol"},"filter_config":{"anyOf":[{"$ref":"#/components/schemas/UniverseFilterConfigSchema"},{"type":"null"}]}},"type":"object","required":["start_date","end_date"],"title":"GenerateRankingsRequest"},"GenerateRankingsResponse":{"properties":{"universe_code":{"type":"string","title":"Universe Code"},"strategy_name":{"type":"string","title":"Strategy Name"},"strategy_version":{"type":"string","title":"Strategy Version"},"benchmark_symbol":{"type":"string","title":"Benchmark Symbol"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"trading_days_total":{"type":"integer","title":"Trading Days Total"},"runs_created":{"type":"integer","title":"Runs Created"},"runs_reused":{"type":"integer","title":"Runs Reused"},"runs_failed":{"type":"integer","title":"Runs Failed"},"failed_dates":{"items":{"type":"string","format":"date"},"type":"array","title":"Failed Dates"}},"type":"object","required":["universe_code","strategy_name","strategy_version","benchmark_symbol","start_date","end_date","trading_days_total","runs_created","runs_reused","runs_failed"],"title":"GenerateRankingsResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IngestBatchStatus":{"type":"string","enum":["success","partial_success"],"title":"IngestBatchStatus"},"IngestPeriod":{"type":"string","enum":["1mo","3mo","6mo","1y","5y"],"title":"IngestPeriod"},"IngestionRunRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"symbol":{"type":"string","title":"Symbol"},"provider":{"type":"string","title":"Provider"},"requested_period":{"type":"string","title":"Requested Period"},"rows_inserted":{"type":"integer","title":"Rows Inserted"},"rows_updated":{"type":"integer","title":"Rows Updated"},"rows_skipped":{"type":"integer","title":"Rows Skipped"},"status":{"$ref":"#/components/schemas/IngestionRunStatus"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["id","symbol","provider","requested_period","rows_inserted","rows_updated","rows_skipped","status"],"title":"IngestionRunRead"},"IngestionRunStatus":{"type":"string","enum":["RUNNING","COMPLETED","FAILED"],"title":"IngestionRunStatus"},"MarketDataIngestRequest":{"properties":{"symbols":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Symbols"},"period":{"$ref":"#/components/schemas/IngestPeriod","default":"1y"}},"type":"object","required":["symbols"],"title":"MarketDataIngestRequest"},"MarketDataIngestResponse":{"properties":{"symbols_processed":{"type":"integer","title":"Symbols Processed"},"symbols_failed":{"type":"integer","title":"Symbols Failed"},"rows_inserted":{"type":"integer","title":"Rows Inserted"},"rows_updated":{"type":"integer","title":"Rows Updated"},"rows_skipped":{"type":"integer","title":"Rows Skipped"},"status":{"$ref":"#/components/schemas/IngestBatchStatus"},"runs":{"items":{"$ref":"#/components/schemas/IngestionRunRead"},"type":"array","title":"Runs"}},"type":"object","required":["symbols_processed","symbols_failed","rows_inserted","rows_updated","rows_skipped","status"],"title":"MarketDataIngestResponse"},"MarketDataRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"stock_id":{"type":"string","format":"uuid","title":"Stock Id"},"date":{"type":"string","format":"date","title":"Date"},"open":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Open"},"high":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"High"},"low":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Low"},"close":{"type":"number","title":"Close"},"volume":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Volume"},"adj_close":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Adj Close"},"dividend":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Dividend"},"split_factor":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Split Factor"},"source":{"type":"string","title":"Source"},"ingested_at":{"type":"string","format":"date-time","title":"Ingested At"}},"type":"object","required":["id","stock_id","date","open","high","low","close","volume","adj_close","dividend","split_factor","source","ingested_at"],"title":"MarketDataRead"},"RankingResultRead":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"stock_id":{"type":"string","title":"Stock Id"},"symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Symbol"},"rank":{"type":"integer","title":"Rank"},"score":{"type":"number","title":"Score"},"score_components":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Score Components"}},"type":"object","required":["stock_id","rank","score"],"title":"RankingResultRead"},"RankingRunRead":{"properties":{"id":{"type":"string","title":"Id"},"universe_code":{"type":"string","title":"Universe Code"},"as_of_date":{"type":"string","format":"date","title":"As Of Date"},"strategy_name":{"type":"string","title":"Strategy Name"},"strategy_version":{"type":"string","title":"Strategy Version"},"benchmark_symbol":{"type":"string","title":"Benchmark Symbol"},"inputs_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Inputs Hash"},"filter_config_hash":{"type":"string","title":"Filter Config Hash"},"normalization_method":{"type":"string","title":"Normalization Method"},"status":{"type":"string","title":"Status"},"started_at":{"type":"string","title":"Started At"},"completed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Completed At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"},"results_count":{"type":"integer","title":"Results Count","default":0},"results":{"items":{"$ref":"#/components/schemas/RankingResultRead"},"type":"array","title":"Results"}},"type":"object","required":["id","universe_code","as_of_date","strategy_name","strategy_version","benchmark_symbol","filter_config_hash","normalization_method","status","started_at"],"title":"RankingRunRead"},"RankingRunRequest":{"properties":{"universe_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Universe Code"},"as_of_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"As Of Date"},"strategy_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Name"},"strategy_version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy Version"},"benchmark_symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Benchmark Symbol"},"filter_config":{"anyOf":[{"$ref":"#/components/schemas/UniverseFilterConfigSchema"},{"type":"null"}]}},"type":"object","title":"RankingRunRequest"},"RankingTopRead":{"properties":{"run_id":{"type":"string","title":"Run Id"},"as_of_date":{"type":"string","format":"date","title":"As Of Date"},"strategy_name":{"type":"string","title":"Strategy Name"},"strategy_version":{"type":"string","title":"Strategy Version"},"top":{"items":{"$ref":"#/components/schemas/RankingResultRead"},"type":"array","title":"Top"}},"type":"object","required":["run_id","as_of_date","strategy_name","strategy_version","top"],"title":"RankingTopRead"},"RegimeIcRead":{"properties":{"bull_low_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bull Low Vol Ic"},"bull_high_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bull High Vol Ic"},"bear_low_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bear Low Vol Ic"},"bear_high_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bear High Vol Ic"}},"type":"object","title":"RegimeIcRead"},"StockRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"symbol":{"type":"string","title":"Symbol"},"name":{"type":"string","title":"Name"},"exchange":{"type":"string","title":"Exchange"},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"industry":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Industry"},"is_active":{"type":"boolean","title":"Is Active"},"data_status":{"type":"string","title":"Data Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","symbol","name","exchange","sector","industry","is_active","data_status","created_at","updated_at"],"title":"StockRead"},"UniverseFilterConfigSchema":{"properties":{"min_history_days":{"type":"integer","minimum":1.0,"title":"Min History Days","default":63},"min_avg_daily_traded_value":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Min Avg Daily Traded Value","default":"10000000"},"min_stock_price":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Min Stock Price","default":"50"},"require_data_status_active":{"type":"boolean","title":"Require Data Status Active","default":true},"require_stock_active":{"type":"boolean","title":"Require Stock Active","default":true},"market_data_source":{"type":"string","title":"Market Data Source","default":"yahoo"}},"type":"object","title":"UniverseFilterConfigSchema"},"ValidationBackfillRequest":{"properties":{"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"force_recompute":{"type":"boolean","title":"Force Recompute","default":false}},"type":"object","required":["start_date","end_date"],"title":"ValidationBackfillRequest"},"ValidationBackfillResponse":{"properties":{"runs_found":{"type":"integer","minimum":0.0,"title":"Runs Found"},"validated":{"type":"integer","minimum":0.0,"title":"Validated"},"reused":{"type":"integer","minimum":0.0,"title":"Reused"},"failed":{"type":"integer","minimum":0.0,"title":"Failed"}},"type":"object","required":["runs_found","validated","reused","failed"],"title":"ValidationBackfillResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValidationReportRead":{"properties":{"ranking_run_id":{"type":"string","title":"Ranking Run Id"},"status":{"type":"string","title":"Status"},"validation_hash":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Validation Hash"},"regime_label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Regime Label"},"trend_regime":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Trend Regime"},"vol_regime":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Vol Regime"},"horizon_metrics":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Horizon Metrics"},"sample_summary":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Sample Summary"},"computed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Computed At"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"}},"type":"object","required":["ranking_run_id","status"],"title":"ValidationReportRead"},"ValidationSnapshotRead":{"properties":{"id":{"type":"string","title":"Id"},"stock_id":{"type":"string","title":"Stock Id"},"symbol":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Symbol"},"return_5d":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Return 5D"},"return_10d":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Return 10D"},"return_20d":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Return 20D"},"return_60d":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Return 60D"},"captured_at":{"type":"string","title":"Captured At"}},"type":"object","required":["id","stock_id","captured_at"],"title":"ValidationSnapshotRead"},"ValidationSummaryRead":{"properties":{"reports_count":{"type":"integer","title":"Reports Count"},"horizon":{"type":"integer","title":"Horizon"},"validated_runs":{"type":"integer","title":"Validated Runs","default":0},"failed_runs":{"type":"integer","title":"Failed Runs","default":0},"insufficient_data_runs":{"type":"integer","title":"Insufficient Data Runs","default":0},"average_ic_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Average Ic 20D"},"median_ic_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Median Ic 20D"},"top_decile_return_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Top Decile Return 20D"},"bottom_decile_return_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bottom Decile Return 20D"},"spread_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Spread 20D"},"hit_rate_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Hit Rate 20D"},"directional_hit_rate_20d":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Directional Hit Rate 20D"},"bull_market_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bull Market Ic"},"bear_market_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bear Market Ic"},"high_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"High Vol Ic"},"low_vol_ic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Low Vol Ic"},"regime_ic":{"anyOf":[{"$ref":"#/components/schemas/RegimeIcRead"},{"type":"null"}]},"best_regime":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Best Regime"},"worst_regime":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Worst Regime"}},"type":"object","required":["reports_count","horizon"],"title":"ValidationSummaryRead"}}}}